backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
scikit-learn
psutil
python-jose[cryptography]
passlib[bcrypt]
